NPSS Rline to T-MATS format tool
Tool to create T-MATS format maps from NPSS maps that use Rlines.
Because the NPSS map interface can be changed by a NPSS developer it is
impossible to make a tool that can translate all NPSS maps to T-MATS
maps.

This file shows how this translation can be done and offers a starting
point for the development of a custom NPSS map to T-MATS map translator.