T-MATS Cantera Classes
Updated TMATSC to clean up the interface between TMATSC scripts and
Cantera interface scripts. This includes creating a package "TMATSC" and
a special flow class that is used to integrate with Cantera commands.
The command "import TMATSC.*" can be used to include the package into a
script or the commands can be called directly by typing
TMATSC.FunctionName .

Also updated the verison number to be 1.1.1d for development.  This
should again be changed to 1.1.1 for release.