Bus Update
Update busses to more simple names.  This allows the current T-MATS bus
system to be used with the automatic bus object generator in Simulink.
Making it much easier to wire Station Data out of the Solver area when
using libraries or model blocks.

Also updated all FARc to FAR to be simpler and updated all help docs to
contain all information that was in the busses.