Gear box flip side A and B
Gear Box mask inputs were typically being input as the recipricals of
the Gear Ratio and Efficiency due to frame of reference. This was
leading to some confusion.

To correct this, Side A and B were flipped. Now the Gear Ratio will
typically be greater than one and the efficiency will typically be less
than one.