Turbine Errors
Added warnings to notify a user when the sizes of the map file
definitions are wrong. Also added new outputs and combined the baseline
turbine and the NPSS version. There were also some name changes for mask
inputs (note, this will cause old inputs to be removed when using older
models)