Un-choked update
updated unchoked nozzle to use ideal Pressure expansion.
Added in warning for when nozzle total pressure is equal to or less than
ambient pressure.

This was done to increase the robustness of the code purposes and
consistency.  Additionally it was determined that Area was taken into
account with the flow input / calculated input iteration. This achieves
a more reasonable throat velocity when unchoked.