Add enthalpy input to Burner
Added an enthalpy input to the burner. This gives the user an alternate
way of describing the Fuel.  In the current system hFuel will be equal
to LHV * Eff.  However if the thermodynamic table lookup files are
updated with Cantera the hFuel would need to be changed to include the
heat of formation.