Jacobian Calculator update multi step
Updated Jacobian calculator to have the ability to use more than one
perturbation. This was implemented by adding another dimension onto the
Perturbation definition.  Operation will occur now by running sets of
perturbations, calculating the Jacobian for each set then averaging
them.

It should be noted that if a solver or Jacobian block was created using
a column vector, the new implementation will not understand. For a
single perturbation the vector must now be a row vector.  Additional
perturbations are defined as additional rows.