update variable names in library reference
updated the library reference model to use names in library models. The
reference model was using some old names for the compressor and turbine
blocks that was causing warnings.  MATLAB updated this automatically, it
was just required to open the block, allow MATLAB to get the correct
names, then resave.