A removal
fixed bug having to do with the creation of an A variable in the
workspace. This was tracked down to the nozzle component where it was
being defined in 2 of the mask variable callbacks.

A was being created to hold the values from a get statement for a
switch.  The fix was to use the get statement with the switch directly
thereby eliminating the need for an intermediary variable.

Also decided to implement a third number in the version number to denote
the commit number of the non-released version.