Unchoked nozzle with Ides
updated the nozzle module to calculate the IDes  Area with the throat
conditions instead of MN1 conditions.  The code will produce a warning
when the throat is not choked.

Additionally updated the SWtype to be convergent or convergent-divergent
with a drop down (the old way with a value was prone to error).
Updating the MSK enable script to hand integer inputs was required for
this.