time dependent data plotter
Added a data plotter that will search a model file for output data that
matches the T-MATS data style and then plot it.  the function is located
within the TMATS class  and is called TDplot.

This script will operate on Station data, plot compressor and turbine
maps, and generate speed plots.

after loading type: help TMATS for more info.