Plot function update return fig settings
updated the plot function to return old plot settings.
Also, updated to update only the LibD and gb values when looking at
alternate models.  Performing all functions separately for different
models was creating issues with plotting data on the maps (many times
the ToWorkspace block and compressor or turbine are not in the same
model.