S-function scalar breakout v1.1.3.2
Run turbine and compressor map scalars out of the S-functions and add
constant blocks that will allow a user to use them dynamically more
easily.

A simple customization would be to update the compressor block, removing
the constant block and adding an outport for the efficiency scalar.
Note: this will not remove the variable in the mask (the mask variable
just wont do anything), also this will disable iDesign for the
efficiency scalar (using the iDesign variable overwrites the variable in
the constant block).