CompressorVG v1.1.3.3
Updated compressor block to facilitate the addition of a 3rd dimension
to the performance maps

Created CompressorVG.  This block operates similar to the baseline
compressor, but uses 3d performance maps. the extra dimension is
dedicated to an Alpha vector. This vector can be used for whatever a
user likes, but is typically used for variable geometries.