Heat Soak block v1.1.3.8
Created heat soak block. This block models the effects of a material
(lump mass) on the fluid stream. At a high level the block caluclates
the convective heat transfer between the two mediums, adjusts the energy
of the fluid stream to find the heat soaked temperature, then determines
the change in metal temperature for a dTm/dt to be used to update Tm
(see block help for more information on the subject).