Solver Build sizing and speed v1.1.3.10
This change updates the solvers, JC, and NR in the following ways:
1) Defines sizes of each inport and outport. This helps matlab correctly
determine sizing during a build and cuts back on build errors.
2) Each solver error (from IC, jacobian can't be inverted, and NR, hit
bound limit) was made to display only one time. This cuts back on speed
issues when the system is not converging correctly. Note: MATLAB blocks
are now used for min/max functionality (it was found to be faster than
calling a MATLAB function block to do the same thing).
3) Sizing warning for the NR was fixed, the warning was occuring when
the size was ok. Essencially NR input for IC was changed to
ones(size(Solver_IC)) from ones(size(Solver_IC,2),1).  The solver will
request a mx1 vector for the ICs, but a 1xm vector will work as well
(note, Mx size must match the IC size, i.e. IC as a mx1 and Mx as a 1xm
or visa versa will not work)
4) redefined default values in the library browser to be in line with
current input requirements.