1-D heat Xfer block for all cases v1.1.3.12
Posted 1-D heat Xfer block as a replacement for all previous 1-D heat
Xfer blocks. This block has it all, and is highly adjustable. Boundary
Convection, Conduction, and Radiation are all available. Node sizes and
Conduction coefficients are adjustable as a function of distance.
Material contants can all be adjusted dynamically. This block also has
the options of planar or cylindrical coordinates.