Block Replacer v1.1.3.14
Added script to replace blocks.  This script is designed to be used in
mask function initialization scripts (note: Allow library block to
modify its contents must be checked on) and can be used to swap inports
or outports with contants or terminators respectively to add or remove
block inputs and outputs.