PW Lin Gen robustness v1.1.3.24
Made several updates to increase block robustness.

Added data type convert blocks to Data output. There was a potential for errors when combining boolean and doubles into a bus.

Updated X, Env, and U outports to output last good values when generator is finished. This removes the large ending transient. This transient did not affect functionality, but it looks better without it.

Issues with initialization were causing Env initial outputs to be zero. Proper ICs were placed in memory blocks and the liniearization block was set to use proper ICs before the slew portion of the block has had a chance to run.