Jacobian Inversion fault v1.1.3.28
Added an output from the Jacobian calculator to give a boolean when the Inversion failed.  This is signal is then also output from the Iterative and Steady state solvers in the S_Data signal as JacFailed. The purpose for this is to give a user another handle for shutting the simulation down if the jacobian is failing to solve.