NPSS to T-MATS conversion tool v1.1.3.29
This tool is designed create a T-MATS model from an NPSS model. Using the tool comes in three parts.
1) mapplot_matlab.fnc - script is designed to generate matlab style maps from NPSS maps.
2) BlockTransMTLB_TMATS.fnc - gathers data from the NPSS model that will be required for the T-MATS model.
3) AutoLink_TMATS - AutoGenerates the T-MATS model based on inputs gathered with the previous script.

This code relies on a Like-for-Like model generation methodology and assumes certain NPSS have an equivalent T-MATS component with certain required inputs and also assumes these NPSS components are default. Once the model is generated, similarity with the NPSS model must be validated. At this time, these scripts generate the model only, ind/dep solvers and other modeling requirements must be added once the tool is complete by the user.