Install directory update v1.1.3.32
Update install directory code to be more robust.

The script will now look at the location of the running script for base install file location. This path will the be used as the root to add other paths to MATLAB.

This allows a user to install T-MATS while not in the T-MATS folder.