Tools update v1.1.3.35
Updated SSplot and TDplot to simply try to close open simulink diagrams. This allows the use of calling plotting scripts from within the simulink diagram. This feature will be used for the T-MATS examples.

updated Link_Setup to include the mixer block and a few other blocks. This is needed for the NPSS to T-MATS building tool.